third_party/blink/web_tests/external/wpt/web-animations/animation-model/animation-types/interpolation-per-property-001-expected.txt

This is a testharness.js-based test.
Found 9 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] background-image uses discrete animation when animating between "url("http://localhost/test-1")" and "url("http://localhost/test-2")" with linear easing
  assert_equals: The value should be url("http://localhost/test-1") at 499ms expected "url(\\"http://localhost/test-1\\")" but got "cross-fade(url(\\"http://localhost/test-1\\") 50.1%, url(\\"http://localhost/test-2\\"))"
[FAIL] background-image uses discrete animation when animating between "url("http://localhost/test-1")" and "url("http://localhost/test-2")" with effect easing
  assert_equals: The value should be url("http://localhost/test-1") at 940ms expected "url(\\"http://localhost/test-1\\")" but got "cross-fade(url(\\"http://localhost/test-1\\") 57.1741%, url(\\"http://localhost/test-2\\"))"
[FAIL] background-image uses discrete animation when animating between "url("http://localhost/test-1")" and "url("http://localhost/test-2")" with keyframe easing
  assert_equals: The value should be url("http://localhost/test-1") at 940ms expected "url(\\"http://localhost/test-1\\")" but got "cross-fade(url(\\"http://localhost/test-1\\") 57.1741%, url(\\"http://localhost/test-2\\"))"
[FAIL] border-image-source uses discrete animation when animating between "url("http://localhost/test-1")" and "url("http://localhost/test-2")" with linear easing
  assert_equals: The value should be url("http://localhost/test-1") at 499ms expected "url(\\"http://localhost/test-1\\")" but got "cross-fade(url(\\"http://localhost/test-1\\") 50.1%, url(\\"http://localhost/test-2\\"))"
[FAIL] border-image-source uses discrete animation when animating between "url("http://localhost/test-1")" and "url("http://localhost/test-2")" with effect easing
  assert_equals: The value should be url("http://localhost/test-1") at 940ms expected "url(\\"http://localhost/test-1\\")" but got "cross-fade(url(\\"http://localhost/test-1\\") 57.1741%, url(\\"http://localhost/test-2\\"))"
[FAIL] border-image-source uses discrete animation when animating between "url("http://localhost/test-1")" and "url("http://localhost/test-2")" with keyframe easing
  assert_equals: The value should be url("http://localhost/test-1") at 940ms expected "url(\\"http://localhost/test-1\\")" but got "cross-fade(url(\\"http://localhost/test-1\\") 57.1741%, url(\\"http://localhost/test-2\\"))"
[FAIL] flex-flow uses discrete animation when animating between "row" and "row-reverse wrap" with linear easing
  assert_equals: The value should be row at 0ms expected "row" but got "row nowrap"
[FAIL] flex-flow uses discrete animation when animating between "row" and "row-reverse wrap" with effect easing
  assert_equals: The value should be row at 0ms expected "row" but got "row nowrap"
[FAIL] flex-flow uses discrete animation when animating between "row" and "row-reverse wrap" with keyframe easing
  assert_equals: The value should be row at 0ms expected "row" but got "row nowrap"
Harness: the test ran to completion.
